+++ x/btclightclient/types/expected_keepers.go
@@ -7,9 +7,9 @@ import (
 )
 
 type BTCLightClientHooks interface {
-	AfterBTCRollBack(ctx context.Context, headerInfo *BTCHeaderInfo)       // Must be called after the chain is rolled back
-	AfterBTCRollForward(ctx context.Context, headerInfo *BTCHeaderInfo)    // Must be called after the chain is rolled forward
-	AfterBTCHeaderInserted(ctx context.Context, headerInfo *BTCHeaderInfo) // Must be called after a header is inserted
+	AfterBTCRollBack(ctx context.Context, rollbackFrom, rollbackTo *BTCHeaderInfo) // Must be called after the chain is rolled back
+	AfterBTCRollForward(ctx context.Context, headerInfo *BTCHeaderInfo)            // Must be called after the chain is rolled forward
+	AfterBTCHeaderInserted(ctx context.Context, headerInfo *BTCHeaderInfo)         // Must be called after a header is inserted
 }
 
 type IncentiveKeeper interface {