+++ x/btcstaking/types/keys.go
@@ -24,7 +24,8 @@ var (
 	// 0x05 was used for something else in the past
 	BTCHeightKey = []byte{0x06} // key prefix for the BTC heights
 	// 0x07 was used for something else in the past
-	PowerDistUpdateKey        = []byte{0x08}             // key prefix for power distribution update events
-	AllowedStakingTxHashesKey = collections.NewPrefix(9) // key prefix for allowed staking tx hashes
-	HeightToVersionMapKey     = []byte{0x10}             // key prefix for height to version map
+	PowerDistUpdateKey        = []byte{0x08}              // key prefix for power distribution update events
+	AllowedStakingTxHashesKey = collections.NewPrefix(9)  // key prefix for allowed staking tx hashes
+	HeightToVersionMapKey     = []byte{0x10}              // key prefix for height to version map
+	LargestBtcReorgInBlocks   = collections.NewPrefix(13) // key prefix for the BTC block height difference of the largest reorg
 )